Move greedySensorPositions to frameworks/layers
