+++ htmresearch/frameworks/layers/sensor_placement.py
